+++ entranceSystem/log.txt
@@ -1,0 +1,15 @@
+Log Name: User Added ---- Log Description: A New User Has Been Added  
+Log Name: User %s Added  ---- Log Description: A New User Has Been Added  
+Log Name: User Added ---- Log Description: A New User Has Been Added  
+Log Name: User Added ---- Log Description: A New User Has Been Added ---  07/01/2019, 23:24
+Log Name: Clean Up ---- Log Description: Clean Up Completed ---  07/01/2019, 23:26
+Log Name: Print List ---- Log Description: Print users list ---  07/01/2019, 23:27
+Log Name: User Added ---- Log Description: A New User Has Been Added ---  12/01/2019, 11:44
+Log Name: Print List ---- Log Description: Print users list ---  12/01/2019, 11:45
+Log Name: Print List ---- Log Description: Print users list ---  12/01/2019, 11:45
+Log Name: Print List ---- Log Description: Print users list ---  12/01/2019, 11:45
+Log Name: User Added ---- Log Description: A New User Has Been Added ---  12/01/2019, 11:46
+Log Name: Print List ---- Log Description: Print users list ---  12/01/2019, 11:46
+Log Name: Print List ---- Log Description: Print users list ---  12/01/2019, 11:46
+Log Name: Clean Up ---- Log Description: Clean Up Completed ---  12/01/2019, 11:47
+Log Name: Clean Up ---- Log Description: Clean Up Completed ---  12/01/2019, 11:48
